requirements.txt
streamlit
langchain-community
langchain-google-genai
langchain-huggingface
qdrant-client
uuid
langchain
langchain-core
sentence-transformers
python-dotenv
pymupdf
langchain-qdrant
blinker